airflow/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
azure-core==1.38.2
azure-storage-blob==12.28.0
python-dotenv==1.2.1
lxml==6.0.2